requirements-api.txt
# RentSignal API — deployment dependencies (lean, no notebook/spatial processing deps)

# API
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0

# ML (pinned to match training versions — mismatch causes wrong predictions)
scikit-learn>=1.6,<1.7
xgboost>=3.0,<4.0
shap>=0.50,<1.0
joblib>=1.4,<2.0

# Data (pandas for feature prep, scipy for KDTree spatial queries)
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scipy>=1.13,<2.0
pyarrow>=15.0,<20.0

# Auth (Supabase JWT validation)
PyJWT>=2.8,<3.0

# Database (Supabase client for portfolio CRUD)
supabase>=2.0,<3.0

# HTTP client (OSM Photon address autocomplete)
httpx>=0.27,<1.0

# File uploads (CSV import UploadFile)
python-multipart>=0.0.9

# AI (Gemini image analysis for photo upload)
google-genai>=1.0,<2.0
pillow>=10.0,<12.0

# Utils
orjson>=3.10,<4.0
